Fix LP64 bugs in the pure interpreter; a real ROM now boots past IPL3
A retail cart was dead-ending at 0x800001c8 - BGEZAL r0,-1, IPL3's CIC
checksum failure loop, reached from BNE r7,r8 against the header CRC1/CRC2.

The data was never at fault. Verified byte-exact against the raw file: every
ROMCache_pointer() read including across the 1 MiB window boundary, at 2 MiB
and at 16 MiB, and the entire DMA'd 1 MiB in RDRAM (262,144 words, matching
sum, xor, first and last word). The arithmetic was wrong.

`long` is 32 bits on SH4 and PPC32 but 64 on an LP64 host, and the
interpreter uses `long` as "the 32-bit MIPS word":

- macros.h sign_extended() was (long long)((signed long)a) - a no-op on
  LP64, so 32-bit results were never truncated and registers carried garbage
  in their upper half.
- macros.h rrt32/rrd32/rrs32/irs32/irt32 read and wrote all 64 bits.
- pure_interp.c SRL and SRLV cast to (unsigned long), sign-extending a
  negative operand and then shifting the high 0xFFFFFFFF down into it.
- pure_interp.c DIVU divided that sign-extended value.
- SLL/SLLV/SRA/SRAV were correct only by truncation; made explicit.

All fixed with int/unsigned int, which is 32 bits on SH4, PPC32 and LP64
alike - identical codegen for GC/Wii, correct for the host stub.

None of these would fail on real Dreamcast hardware, where long is 32 bits.
They are host-stub fidelity bugs, and the reason the stub could not run a
real ROM. Not64 on Wii never hit them because it runs the PPC dynarec rather
than pure_interp.c.

Result: both boot checksums now pass, IPL3 jumps to the game, and 600M
instructions later the PC is in game code taking a TLB store miss
(Cause=0x0c, EPC=0x800afbe4). VI origin is still 0, so Phase 4 is not
unblocked - but the failure moved two layers deeper. Phase 3.5 in PORTING.md
has the evidence and points the next owner at the TLB refill path.

Test coverage, because these survived exactly by not being tested:
- CPUTEST now runs SRL/SRLV/SRA/SRAV/SLLV/SUBU/DIVU/MFLO/MFHI on a NEGATIVE
  operand (0x95F208B7). A positive one passes with or without the bugs.
- It checks one full 64-bit register (r22 = 0xFFFFFFFFF208B700). The low half
  looks correct even when the high half is garbage, which is how the
  macros.h fault survived every low-32 check; IPL3 compares 64-bit registers
  with SLTU, so the high half is load-bearing.
- Each fix was verified to fail CPUTEST when individually reverted.

Makefile.dc now uses -MMD -MP instead of a hand-written dep list. That list
omitted macros.h, so editing it silently reused stale objects and produced
runs that disagreed with the source - which cost real time during this
investigation.

make -f Makefile.dc HOST=1 test passes under gcc and clang.

+++ r4300/pure_interp.c
@@ -82,42 +82,42 @@ static void NI()
 
 static void SLL()
 {
-   rrd32 = (unsigned long)(rrt32) << rsa;
+   rrd32 = (unsigned int)(rrt32) << rsa;
    sign_extended(rrd);
    interp_addr+=4;
 }
 
 static void SRL()
 {
-   rrd32 = (unsigned long)rrt32 >> rsa;
+   rrd32 = (unsigned int)rrt32 >> rsa;
    sign_extended(rrd);
    interp_addr+=4;
 }
 
 static void SRA()
 {
-   rrd32 = (signed long)rrt32 >> rsa;
+   rrd32 = (signed int)rrt32 >> rsa;
    sign_extended(rrd);
    interp_addr+=4;
 }
 
 static void SLLV()
 {
-   rrd32 = (unsigned long)(rrt32) << (rrs32&0x1F);
+   rrd32 = (unsigned int)(rrt32) << (rrs32&0x1F);
    sign_extended(rrd);
    interp_addr+=4;
 }
 
 static void SRLV()
 {
-   rrd32 = (unsigned long)rrt32 >> (rrs32 & 0x1F);
+   rrd32 = (unsigned int)rrt32 >> (rrs32 & 0x1F);
    sign_extended(rrd);
    interp_addr+=4;
 }
 
 static void SRAV()
 {
-   rrd32 = (signed long)rrt32 >> (rrs32 & 0x1F);
+   rrd32 = (signed int)rrt32 >> (rrs32 & 0x1F);
    sign_extended(rrd);
    interp_addr+=4;
 }
@@ -272,8 +272,8 @@ static void DIVU()
 {
    if (rrt32)
      {
-	lo = (unsigned long)rrs32 / (unsigned long)rrt32;
-	hi = (unsigned long)rrs32 % (unsigned long)rrt32;
+	lo = (unsigned int)rrs32 / (unsigned int)rrt32;
+	hi = (unsigned int)rrs32 % (unsigned int)rrt32;
 	sign_extended(lo);
 	sign_extended(hi);
      }
